requirements.txt
asgiref==3.10.0
Django==5.2.7
django-login-as==0.2
pillow==12.0.0
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0